GoRouterを使わずにリファクタリング
